requirements.txt
setuptools==68.0.0
requests==2.31.0
charset-normalizer==3.3.2
idna==3.11
urllib3==2.1.0
certifi==2026.1.4
python-dotenv==1.0.0
